% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizePosterior.R
\name{BIC.mix}
\alias{BIC.mix}
\title{Calculate the BIC (information-theoretic criteria) for a mixture model.}
\usage{
\method{BIC}{mix}(out)
}
\arguments{
\item{l1}{Log likelihood 1}

\item{l2}{Log likelihood 1}

\item{n}{Number of total tests for Benfferroni correction}
}
\description{
This function calculate the BIC (information-theoretic criteria) for a mixture model.
}
\examples{
detectVarGenes(suva_exp,500)
}
\keyword{Detect}
\keyword{genes}
\keyword{varibale}
