% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{calcNormFactors}
\alias{calcNormFactors}
\title{Calculate a normalization factor for each column in an expression matrix}
\usage{
calcNormFactors(expmat)
}
\arguments{
\item{expmat}{A genes X samples expression matrix scaled to Log2(CPM/10+1).}
}
\description{
This function calculates an adjustment factor for each sample in a log2(CPM/10+1)-scaled expression matrix.
}
\examples{
calcNormFactors(suva_exp)
}
\keyword{Adjustment}
\keyword{factor}
