% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarizeMatrix.R
\name{calcPvalue}
\alias{calcPvalue}
\title{Calculate p-values from a mixmdl object.}
\usage{
calcPvalue(mixmdl, normal, tumor, threshold = 0.8)
}
\arguments{
\item{mixmdl}{An mixmdl object returned by the mixtools package.}

\item{normal}{Vector of positions indicating the indices that identify posteriors assigned to normal cells.}

\item{tumor}{Vector of positions indicating the indices that identify posteriors assigned to tumor cells.}

\item{threshold}{Optional: A threshold for binarizing the matrix. Example: For gains all cells with a posterior probability >threshold for the component with the larger mean will be assigned a 1.}
}
\description{
This function generates p-values from a mixmdl object based on the mean and variance of the two components
}
\examples{
calcPvalue (mixmdl,normal,tumor)
}
\keyword{Binarize}
\keyword{vector}
