% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tsne.R
\name{calculateTsne}
\alias{calculateTsne}
\title{Calculate a 2D t-Distributed Stochastic Neighbor Embedding from a gene expression matrix}
\usage{
calculateTsne(mat, genes, dims = 20, perplexity = 30)
}
\arguments{
\item{genes}{A list of genes used for tSNE generation (HUGO format)}

\item{expmat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}
}
\description{
This function calculates a 2D t-Distributed Stochastic Neighbor Embedding from a gene expression matrix in the space of a given set of genes
}
\examples{
calculateTsne(suva_exp,c("EGFR","PDGFRA","MET","SOX2"))
}
\keyword{Plot}
\keyword{tSNE}
