% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferBreakpoints.R
\name{detectBreakPoints}
\alias{detectBreakPoints}
\title{Visualize normal/tumor ratio of genes expressed from the same chromosome and sorted by position}
\usage{
detectBreakPoints(mat, normal, tumor, windowsize, gene_pos, chr, patients,
  patient, breakpoints)
}
\arguments{
\item{mat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}

\item{normal}{Indices of columns holding normal cells.}

\item{tumor}{Indices of columns holding tumor cells.}

\item{windowsize}{Size of the window smoothing the gene expression data}

\item{gene_pos}{Matrix with positional information for each gene}

\item{chr}{Chromosome to plot}

\item{patients}{Vector of patient identifiers for each cell}

\item{patient}{Id of the patient to visualize}

\item{breakpoints}{Chromosomal positions of chromosome arms}
}
\description{
This function allows to visualize the smoothed normal/tumor ratio of genes expressed from the same chromosome and sorted by position. This give information on breakpoints as well as CNV status (gain/loss)
}
\examples{
detectBreakPoints (suva_exp,normal,tumor,101,gene_pos,1,patients,patient,breakpoints)
}
