% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizePosterior.R
\name{extractSlidingRegionsFromGenes}
\alias{extractSlidingRegionsFromGenes}
\title{Extract coordinats of a sliding window across expressed genes}
\usage{
extractSlidingRegionsFromGenes(mat, gene_pos, fracChrGenesInWindow = 0.1,
  windowSlideFraction = 0.25, minWindowSize = 100)
}
\arguments{
\item{mat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}

\item{gene_pos}{A matrix of positions for every gene is the expression matrix, calculated via getGenePositions()}

\item{fracChrGenesInWindow}{A fraction indicating the fraction of genes included in each window}

\item{windowSlideFraction}{A fraction indicating the number of genes to move the sliding window as a fraction of the window size (defined via fracChrGenesInWindow)}

\item{minWindowSize}{A minimum number of genes in each region.}
}
\description{
This function extract coordinats of a sliding window across expressed genes. It uses the matrix supplied to \'mat\'
to extract expressed genes and then generates the regions corresponding a sliding window across the genome
(defined by fracChrGenesInWindow and windowSlideFraction) which can directly be used with the \code{plotAll} function.
}
\examples{
detectVarGenes(suva_exp,500)
}
\keyword{Get}
\keyword{regions}
