% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{filterMatrix}
\alias{filterMatrix}
\title{Filter and expression matrix}
\usage{
filterMatrix(mat1, genelist, minExp = 10, minCells = 5)
}
\arguments{
\item{mat1}{A genes X samples expression matrix.}

\item{genelist}{Genes not in this list will be removed from the matrix.}

\item{minExp}{Only genes with rowSums > minExp are being kept. Default:10}

\item{minCells}{Only genes expressed in > minCells samples will be kept}
}
\description{
This function filters an expression matrix by removing genes which are expressed at low levels and in only very few samples.
}
\examples{
filterMatrix(suva_exp, chr_genes)
}
\keyword{Filter}
\keyword{matrix}
