% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarizeMatrix.R
\name{generatePvalMat}
\alias{generatePvalMat}
\title{Calculate p-values from a mixmdl object.}
\usage{
generatePvalMat(mat, regions, normfactor, normal, tumor, gene_pos,
  threshold = 0.8)
}
\arguments{
\item{mat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}

\item{regions}{A matrix of regions to be evaluated}

\item{normal}{Vector of positions indicating the indices that identify posteriors assigned to normal cells.}

\item{tumor}{Vector of positions indicating the indices that identify posteriors assigned to tumor cells.}

\item{gene_pos}{A matrix of positions for every gene is the expression matrix, calculated via getGenePositions()}

\item{threshold}{Optional: A threshold for binarizing the matrix. Example: For gains all cells with a posterior probability >threshold for the component with the larger mean will be assigned a 1.}

\item{normFactor}{A vector of normalization factors for each cell calculated with calcNormFactors()}
}
\description{
This function generates p-values from a log2(CPM/10+1) scaled gene expression matrix using the calcPvalue() function on all the regions given via the regions parameter
}
\examples{
generatePvalMat(suva_expr,regions,normfactor,normal,tumor,gene_pos)
}
\keyword{Binarizee}
\keyword{vector}
