% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPositions.R
\name{getGenePositions}
\alias{getGenePositions}
\title{Receive genomic coordinates of a gene list}
\usage{
getGenePositions(gene_names, ensembl_version = 87)
}
\arguments{
\item{gene_names}{A vector of gene names in HUGO format.}

\item{ensembl_version}{Version of the ENSEMBL database used to quantify gene expression data. Defaul: v87.}
}
\description{
This function allows to receive the genomic positions of a vector of genes in HUGO format.
}
\examples{
getGenePositions(gene_names=c("EGFR","PDGFRA"))
}
\keyword{Chromosomal}
\keyword{positions}
