% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizePosterior.R
\name{likelihoodRatioTest}
\alias{likelihoodRatioTest}
\title{Perform likelihood ratio test given two log likelihoods and a integer numbe rindicating the number of tests performed for multiple testing correction}
\usage{
likelihoodRatioTest(l1, l2, n)
}
\arguments{
\item{l1}{Log likelihood 1}

\item{l2}{Log likelihood 1}

\item{n}{Number of total tests for Benfferroni correction}
}
\description{
This function performs a likelihood ratio test given two log likelihoods and a integer numbe rindicating the number of tests performed for multiple testing correction
}
\examples{
detectVarGenes(suva_exp,500)
}
\keyword{Detect}
\keyword{genes}
\keyword{varibale}
