% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoReceiveData.R
\name{loadPublicData}
\alias{loadPublicData}
\title{Receive a tab-separated expression matrix in gz format from the web.}
\usage{
loadPublicData(weblink = NULL, cancer_id = NULL)
}
\arguments{
\item{weblink}{.}

\item{ensembl_version}{Version of the ENSEMBL database used to quantify gene expression data. Defaul: v87.}
}
\description{
This function allows to receive a tab-separated expression matrix in gz format from the web, for example TCGA RNA-seq expression data from the ucsc xenabrowser.
}
\examples{
getGenePositions(gene_names=c("EGFR","PDGFRA"))
}
\keyword{Chromosomal}
\keyword{positions}
