% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{normMat}
\alias{normMat}
\title{Normalize a expression matrix of raw counts to log2 (CPM/10+1) values}
\usage{
normMat(expmat)
}
\arguments{
\item{expmat}{A genes X samples expression matrix of raw (single cell) RNA-seq counts.}
}
\description{
This function normlizes a matrix of raw gene counts to log2(CPM/10+1).
}
\examples{
normMat(suva_exp)
}
\keyword{Normalize}
