% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferBreakpoints.R
\name{pdist}
\alias{pdist}
\title{Fast calculation of a distance matrix from a gene expression matrix (R dist function is slow)}
\usage{
pdist(tmat)
}
\arguments{
\item{mat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}
}
\description{
This function calculates a distance matrix from a gene expression matrix (R dist function is slow)
}
\examples{
pdist (suva_expr)
}
