% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizePosterior.R
\name{plotAll}
\alias{plotAll}
\title{Estimate and visualize a two-component Gaussian Mixture Model based on a log2(CPM/10+1) expression matrix}
\usage{
plotAll(mat, normFactor, regions, gene_pos, fname, normal = NULL,
  tumor = NULL, postProb = 0.95, repetitions = 5)
}
\arguments{
\item{mat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}

\item{normFactor}{A vector of normalization factors for each cell calculated with calcNormFactors()}

\item{regions}{A matrix of regions to be evaluated}

\item{gene_pos}{A matrix of positions for every gene is the expression matrix, calculated via getGenePositions()}

\item{fname}{A string indicating the name of the output pdf}

\item{normal}{Optional: A vector of indices for normal cells}

\item{tumor}{Optional: A vector of indices for tumor cells}
}
\description{
This function estimates a two component Gaussian Mixture Model on log2(CPM/10+1) expression matrix and visualizes the results.
}
\examples{
detectVarGenes(suva_exp,500)
}
\keyword{Detect}
\keyword{genes}
\keyword{varibale}
