% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferBreakpoints.R
\name{plotAllChromosomes}
\alias{plotAllChromosomes}
\title{Applies detectBreakPoints() for all chromosomes and writes a pdf file with figures}
\usage{
plotAllChromosomes(mat, normal, tumor, windowsize, gene_pos, fname,
  patients = NULL, patient = NULL, breakpoints = NULL, offs = 40)
}
\arguments{
\item{mat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}

\item{normal}{Indices of columns holding normal cells.}

\item{tumor}{Indices of columns holding tumor cells.}

\item{windowsize}{Size of the window smoothing the gene expression data}

\item{gene_pos}{Matrix with positional information for each gene}

\item{fname}{Name for the output file}

\item{Optional:}{patients Vector of patient identifiers for each cell}

\item{Optional:}{patient Id of the patient to visualize}

\item{Optional:}{breakpoints Chromosomal positions of chromosome arms}
}
\description{
This function applies detectBreakPoints() for all chromosomes and writes a pdf file with figures
}
\examples{
detectBreakPoints (suva_exp,normal,tumor,101,gene_pos,1,patients,patient,breakpoints)
}
