% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizePosterior.R
\name{plotChrEnichment}
\alias{plotChrEnichment}
\title{Estimate a two-component Gaussian Mixture Model based on a log2(CPM/10+1) expression matrix}
\usage{
plotChrEnichment(expmat, chr, normFactor, gene_positions, n, groups1 = NULL,
  groups2 = NULL, start = NULL, end = NULL, k = 2)
}
\arguments{
\item{expmat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}

\item{chr}{Number of the chromosome to be considered.}

\item{normFactor}{A vector of normalization factors for each cell calculated with calcNormFactors()}

\item{gene_positions}{A matrix of positions for every gene in the expression matrix, calculated via getGenePositions()}

\item{groups1}{A vector of indices for normal cells}

\item{groups2}{A vector of indices for tumor cells}

\item{start}{Optional: The start position of the region to be considered}

\item{end}{Optional: The end position of the region to be considered}
}
\description{
This function estimates a two component Gaussian Mixture Model on log2(CPM/10+1) expression matrix.
}
\examples{
detectVarGenes(suva_exp,500)
}
\keyword{Detect}
\keyword{genes}
\keyword{varibale}
