% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizePosterior.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Visualize a heatmap of posterior probabilities of cells for component 1 of Gaussian Mixture Models fit for each user-defined region}
\usage{
plotHistogram(pmat, expmat, clusters, zscoreThreshold = 4, patients = NULL,
  celltypes = NULL)
}
\arguments{
\item{pmat}{Matrix of posterior probabilites, as returned by plotAll().}

\item{expmat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}

\item{clusters}{The number of clusters to cut the dendrogram into}

\item{zscoreThreshold}{A threshold on the row-wise z-scores for visualiztion. Default: 4}

\item{patients}{Optional: A vector naming the patient for each cell}

\item{celltypes}{Optional: A vector naming a putative celltype for each cell}

\item{gene_pos}{A matrix of positions for every gene is the expression matrix, calculated via getGenePositions()}
}
\description{
This function allows to plot a heatmap of posterior probabilities of cells for component 1 of Gaussian Mixture Models fit for each user-defined region.
}
\examples{
plotHistogram(pmat,suva_expr,2,zscoreThreshold=4)
}
\keyword{Heatmap}
\keyword{of}
\keyword{posteriors}
