% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tsne.R
\name{plotTsneGene}
\alias{plotTsneGene}
\title{Visualize the expression of a set of genes in tSNE space}
\usage{
plotTsneGene(rt, mat, genes, ncols = 2)
}
\arguments{
\item{rt}{An object returned by calculateTsne() holding tSNE coordinates}

\item{mat}{A log2(CPM+1/10) scaled expression matrix}

\item{genes}{A vector of genes to visualize}
}
\description{
This function plots the 1st and 2nd tSNE coordinates of cells and visualized the expression of a given set of genes
}
\examples{
plotTsneGene(rt,suva_exp,c("EGFR","PDGFRA","MET","SOX2"))
}
\keyword{Plot}
\keyword{tSNE}
