% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tsne.R
\name{plotTsneProbabilities}
\alias{plotTsneProbabilities}
\title{Visualize the posterior probabilities for one component of a gaussian mixture model for each cell in tSNE space}
\usage{
plotTsneProbabilities(rt, mat, prob, pname, ncols = 2)
}
\arguments{
\item{rt}{An object returned by calculateTsne() holding tSNE coordinates}

\item{mat}{A log2(CPM+1/10) scaled expression matrix}

\item{genes}{A vector of posterior probablities to visualize}
}
\description{
This function plots the 1st and 2nd tSNE coordinates of cells and visualizes the posterior probabilities for one component of a gaussian mixture model for each cell
}
\examples{
plotTsneProbabilities(rt,suva_exp,p1)
}
\keyword{Plot}
\keyword{tSNE}
