% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferBreakpoints.R
\name{removeLogScale}
\alias{removeLogScale}
\title{Transforms data matrix from log2(CPM/10+1) to CPM}
\usage{
removeLogScale(expmat)
}
\arguments{
\item{expmat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}
}
\description{
This function transforms counts in a data matrix from log2(CPM/10+1) to CPM
}
\examples{
removeLogScale(suva_exp)
}
\keyword{Transform}
