% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferBreakpoints.R
\name{string.to.colors}
\alias{string.to.colors}
\title{Maps stings to colors}
\usage{
string.to.colors(st)
}
\arguments{
\item{st}{A vector of strings.}
}
\description{
This function transforms a vector of strings, eg. the patient ID for each cell, to colors
}
\examples{
removeLogScale(suva_exp)
}
\keyword{Transform}
