% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizePosterior.R
\name{testWrapper}
\alias{testWrapper}
\title{Test difference of the z-score distribution in the estimated mixed gausian model}
\usage{
testWrapper(set1, set2)
}
\arguments{
\item{set1}{the z-score values of the first estimated gausian model.}

\item{set1}{the z-score values of the second estimated gausian model.}
}
\description{
Test difference of the z-score distribution in the estimated mixed gausian model
both using mean and median as well as the summary statistics. Not meant to be used by the end user
}
\examples{
detectVarGenes(suva_exp,500)
}
\keyword{NA}
