% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarizeMatrix.R
\name{binarizeMatrix}
\alias{binarizeMatrix}
\title{Binarize a matrix of posterior probabilities.}
\usage{
binarizeMatrix(mixmat, normal, tumor, threshold = 0.8, withna = T)
}
\arguments{
\item{normal}{Vector of positions indicating the indices that identify posteriors assigned to normal cells.}

\item{tumor}{Vector of positions indicating the indices that identify posteriors assigned to tumor cells.}

\item{threshold}{Posterior threshold level. The presence of a CNV is assigned to a cell if its posterior exceeds the threshold.}

\item{withna}{Should posteriors that can't be assigned to component 1, but also not to component2, be set to 0 or NA (defualt is set to NA).}

\item{mixmdl}{Vector of posterior probabilities between 0 and 1.}
}
\description{
This function calls binarizeCalls() on all rows of a matrix
}
\examples{
binarizeCalls(mixmdl,normal,tumor,threshold,withna=T)
}
\keyword{Binarizee}
\keyword{vector}
