% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinarizeMatrix.R
\name{calcPvalue}
\alias{calcPvalue}
\title{Calculate p-values from a mixmdl object.}
\usage{
calcPvalue(mixmdl, normal, tumor, threshold = 0.8)
}
\arguments{
\item{normal}{Vector of positions indicating the indices that identify posteriors assigned to normal cells.}

\item{tumor}{Vector of positions indicating the indices that identify posteriors assigned to tumor cells.}

\item{mati}{A cells X regions matrix}

\item{patients}{A vector of length(nrow(mati)) indicating the patient for each cell.}

\item{patient}{Optional: Which patient should the matrix be plotted for.}
}
\description{
This function visualizes a matrix of binary CNV assignment. A 1 indicates the presence, a 0 the absence of a CNV
}
\examples{
plotBinaryMat(mati,patients,normal,tumor,patient="MGH96")
}
\keyword{Binarizee}
\keyword{vector}
