% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferBreakpoints.R
\name{plotChromosomeHeatmap}
\alias{plotChromosomeHeatmap}
\title{Generates a heatmap of gene expression across the genome for each cell}
\usage{
plotChromosomeHeatmap(mat, normal, tumor, windowsize, chr = NULL,
  expThresh = 0.4, thresh = 1)
}
\arguments{
\item{mat}{A genes X samples expression matrix of log2(CPM/10+1) scaled (single cell) RNA-seq counts.}

\item{normal}{Indices of columns holding normal cells.}

\item{tumor}{Indices of columns holding tumor cells.}

\item{windowsize}{Size of the window smoothing the gene expression data}

\item{chr}{Optional: Order the heatmap based on the expression from this chromosome}
}
\description{
This function generates a heatmap of gene expression across the genome for each cell
}
\examples{
plotChromosomeHeatmap (suva_expr,normal,tumor,121,chr=11)
}
